+++ src/components/_private/Flyout/FlyoutTrigger.tsx
@@ -1,6 +1,8 @@
 "use client";
 
 import React from "react";
+import useHotkeys from "hooks/useHotkeys";
+import * as keys from "constants/keys";
 import type * as T from "./Flyout.types";
 import { useFlyoutContext } from "./Flyout.context";
 
@@ -16,27 +18,50 @@ const FlyoutTrigger = (props: T.TriggerProps) => {
 		handleMouseEnter,
 		handleMouseLeave,
 		handleClick,
+		handleOpen,
 		trapFocusMode,
 	} = useFlyoutContext();
 
+	useHotkeys(
+		{
+			[`${keys.UP},${keys.DOWN}`]: () => {
+				if (flyout.status !== "idle") return;
+				handleOpen();
+			},
+		},
+		[handleOpen, flyout.status],
+		{ ref: triggerElRef }
+	);
+
 	let childrenAttributes: Partial<T.TriggerAttributes> = {
-		onClick: handleClick,
 		onBlur: handleBlur,
 		ref: triggerElRef,
 	};
+
+	if (triggerType === "click") {
+		childrenAttributes.onClick = handleClick;
+	}
 
 	if (triggerType === "hover") {
 		childrenAttributes.onMouseEnter = handleMouseEnter;
 		childrenAttributes.onMouseLeave = handleMouseLeave;
 	}
 
-	if (triggerType === "hover" && trapFocusMode !== "action-menu") {
+	if ((triggerType === "hover" && trapFocusMode !== "action-menu") || triggerType === "focus") {
 		childrenAttributes.onFocus = handleFocus;
 		childrenAttributes["aria-describedby"] = id;
 	}
 
-	if (triggerType === "click" || trapFocusMode === "action-menu") {
-		childrenAttributes["aria-haspopup"] = trapFocusMode === "dialog" ? "dialog" : "menu";
+	if (triggerType === "click" || triggerType === "focus" || trapFocusMode === "action-menu") {
+		if (trapFocusMode === "dialog") {
+			childrenAttributes["aria-haspopup"] = "dialog";
+		} else if (trapFocusMode === "selection-menu") {
+			childrenAttributes["aria-haspopup"] = "listbox";
+			childrenAttributes["aria-autocomplete"] = "list";
+		} else {
+			childrenAttributes["aria-haspopup"] = "menu";
+		}
+
 		childrenAttributes["aria-expanded"] = flyout.status !== "idle";
 		childrenAttributes["aria-controls"] = flyout.status !== "idle" ? id : undefined;
 	}