+++ src/constants/attributes.ts
@@ -1,1 +1,2 @@
 export const keyboardModeAttribute = "data-rs-keyboard";
+export const pseudoFocusAttribute = "data-rs-focus";